rename: unified file names in plural
